requirements.txt
dj-database-url==0.5.0
Django==3.2.9
gunicorn==20.1.0
wheel==0.36.2
whitenoise==5.3.0
pytest==7.4.3
pytest-django==4.4.0
django-excel-response
six